Assimilate `lp-hal-procmacros` into `esp-hal-procmacros`
