+++ esp-hal-procmacros/Cargo.toml
@@ -1,10 +1,6 @@
 [package]
-name    = "esp-hal-procmacros"
-version = "0.6.1"
-authors = [
-    "Jesse Braham <jesse@beta7.io>",
-    "Björn Quentin <bjoern.quentin@mobile-j.de>",
-]
+name         = "esp-hal-procmacros"
+version      = "0.6.1"
 edition      = "2021"
 rust-version = "1.67.0"
 description  = "Procedural macros for ESP-HAL"
@@ -12,32 +8,37 @@ repository   = "https://github.com/esp-rs/esp-hal"
 license      = "MIT OR Apache-2.0"
 
 [package.metadata.docs.rs]
-features = ["esp32c3", "interrupt"]
+features = ["esp32c3", "interrupt", "ram"]
 
 [lib]
 proc-macro = true
 
 [dependencies]
 darling          = "0.20.3"
-proc-macro-crate = "1.3.1"
-proc-macro-error = "1.0.4"
-proc-macro2      = "1.0.66"
-quote            = "1.0.33"
-syn              = {version = "2.0.31", features = ["extra-traits", "full"]}
-object           = {version = "0.32.1", optional = true}
 litrs            = "0.4.0"
-# toml_edit is a dependency of proc-macro-crate. Unfortunately they raised their MSRV on 0.19.15 so we pin the version for now
-toml_edit        = "=0.19.14"
+object           = { version = "0.32.1", optional = true }
+proc-macro-crate = "2.0.0"
+proc-macro-error = "1.0.4"
+proc-macro2      = "1.0.69"
+quote            = "1.0.33"
+syn              = { version = "2.0.38", features = ["extra-traits", "full"] }
 
 [features]
-esp32     = []
-esp32c2   = []
-esp32c3   = []
-esp32c6   = ["dep:object"]
-esp32h2   = []
-esp32s2   = ["dep:object"]
-esp32s3   = ["dep:object"]
+# Select a target device:
+esp32       = []
+esp32c2     = []
+esp32c3     = []
+esp32c6     = ["dep:object"]
+esp32c6-lp  = []
+esp32h2     = []
+esp32s2     = ["dep:object"]
+esp32s2-ulp = []
+esp32s3     = ["dep:object"]
+esp32s3-ulp = []
 
-interrupt = []
-rtc_slow  = []
-embassy   = []
+# Gated features:
+embassy       = []
+enum-dispatch = []
+interrupt     = []
+ram           = []
+rtc_slow      = []
