+++ esp32c6-lp-hal/Cargo.toml
@@ -24,7 +24,7 @@ categories = [
 critical-section = { version = "1.1.2", features = ["restore-state-u8"] }
 embedded-hal     = { version = "0.2.7", features = ["unproven"] }
 esp32c6-lp       = { git = "https://github.com/esp-rs/esp-pacs", rev = "a9cad5e", features = ["critical-section"] }
-lp-hal-procmacros = { path = "../lp-hal-procmacros", features = ["esp32c6"] }
+procmacros       = { package = "esp-hal-procmacros", path = "../esp-hal-procmacros", features = ["esp32c6-lp"] }
 riscv            = "0.10.1"
 paste            = "1.0.14"
