+++ ulp-riscv-hal/Cargo.toml
@@ -23,10 +23,10 @@ categories = [
 
 [dependencies]
 embedded-hal      = { version = "0.2.7", features = ["unproven"] }
-lp-hal-procmacros = { path = "../lp-hal-procmacros" }
+procmacros        = { package = "esp-hal-procmacros", path = "../esp-hal-procmacros" }
 paste             = "1.0.14"
-esp32s2-ulp    = { git = "https://github.com/esp-rs/esp-pacs", package = "esp32s2-ulp", optional = true }
-esp32s3-ulp    = { git = "https://github.com/esp-rs/esp-pacs", package = "esp32s3-ulp", optional = true }
+esp32s2-ulp       = { git = "https://github.com/esp-rs/esp-pacs", package = "esp32s2-ulp", optional = true }
+esp32s3-ulp       = { git = "https://github.com/esp-rs/esp-pacs", package = "esp32s3-ulp", optional = true }
 
 
 [dev-dependencies]
@@ -35,5 +35,5 @@ panic-halt       = "0.2.0"
 [features]
 default = []
 debug   = []
-esp32s2 = [ "dep:esp32s2-ulp", "lp-hal-procmacros/esp32s2" ]
-esp32s3 = [ "dep:esp32s3-ulp", "lp-hal-procmacros/esp32s3" ]
+esp32s2 = [ "dep:esp32s2-ulp", "procmacros/esp32s2-ulp" ]
+esp32s3 = [ "dep:esp32s3-ulp", "procmacros/esp32s3-ulp" ]
